% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmesh_mesh.R
\name{dmesh_mesh}
\alias{dmesh_mesh}
\title{From INLA mesh to dual mesh}
\usage{
dmesh_mesh(mesh)
}
\arguments{
\item{mesh}{An INLA mesh object}
}
\value{
An sf MULTIPOLYGON object with the dual mesh polygons
}
\description{
This function constructs a dual mesh from the standard INLA mesh.
}
\details{
Parallelization can be enabled by choosing a plan from the future package and the mesh will be splitted in chunks detemined by the number of workers. Don't forget to reset the plan to sequential when done.
}
\examples{

}
