% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmesh_predictors.R
\name{dmesh_predictors}
\alias{dmesh_predictors}
\title{Summarize predictors to dual mesh}
\usage{
dmesh_predictors(dmesh, predictors)
}
\arguments{
\item{dmesh}{A dual mesh}

\item{predictors}{A SpatRaster with predictor values. If processing in parallel, the SpatRastre needs to be wrapped using \link{\code{terra::wrap}}.}
}
\value{

}
\description{
This function summarizes predictors to dual mesh. Predictors are assumed to be numerical values.
}
\details{
Parallelization can be enabled by choosing a plan from the future package and the mesh will be splitted in chunks detemined by the number of workers. Don't forget to reset the plan to sequential when done.

When the predictors do not cover the extent of the dual mesh, empty cells will be filled with the value of the nearest neighbour cell.
}
\examples{


}
