% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewlgcp.R
\name{ewlgcp}
\alias{ewlgcp}
\title{Effort-Weighted Log Gaussian Cox Process SDM}
\usage{
ewlgcp(
  formula,
  dmesh,
  effort = TRUE,
  adjust = FALSE,
  buffer = TRUE,
  orthogonal = TRUE,
  prior.beta = NULL,
  prior.range = c(50, 0.1),
  prior.sigma = c(1, 0.1),
  smooth = 3/2,
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x1 + x2}.}

\item{dmesh}{A dual mesh as a \code{MULTIPOLYGONS} \code{sf} object}

\item{effort}{Logical. Whether to adjust the model for effort. Default \code{TRUE}.}

\item{adjust}{Logical. Whether to adjust the effort for being species specific. Default \code{TRUE}.}

\item{buffer}{Logical. Whether to add a fictious effort to help reducing prediction outside of the species range. Default \code{TRUE}.}

\item{orthogonal}{Logical. Whether to make the spatial field orthogonal to the predictors? Default \code{TRUE}.}

\item{prior.beta}{x}

\item{prior.range}{x}

\item{prior.sigma}{x}

\item{smooth}{x}

\item{\dots}{Further arguments to pass to \code{inla}}
}
\value{
A model of class \code{INLA}.
}
\description{
Runs the effort-weighted Log Gaussian Cox Process species distribution model
}
\details{

}
\examples{

add(10, 1)


}
\references{
Simpson, D. Illian, J. B., Lindgren, F. Sørbye, S. H. and Rue, H. 2016. Going off grid: computationally efficient inference for log-Gaussian Cox processes. Biometrika, 103(1): 49-70 \url{https://doi.org/10.1093/biomet/asv064}
}
