% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Construct species distribution maps models}
\usage{
map(model, dmesh, dims, region = NULL, sample = FALSE, nsamples = 100)
}
\arguments{
\item{model}{An object of class \code{ppSpace} or of class \code{uniSpace}.}

\item{dmesh}{The dmesh list}

\item{dims}{A vector of length 2 defining the number of pixels to use as rows and columns to define the map.}

\item{region}{A spatial polygon to isolate the region of interest. If none is given, a map is drawn for the entire region covered by the mesh.}

\item{sample}{Logical. Whether to sample from the posterior distribution of each cell using \code{INLA}'s \code{inla.posterior.sample}. Returns a raster stack of each sample along with other types specified. Default to \code{FALSE}. Currently ignored.}

\item{nsamples}{Integer. Number of samples to draw from the posterior. Ignored if \code{sample = FALSE}.}
}
\value{

}
\description{
Constructs mean, standard deviation, and quantile (0.025, 0.5 and 0.975) maps for models calculated using \code{\link{uniSpace}} and \code{\link{ppSpace}}
}
\details{

}
\references{

}
