% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmesh_effort.R
\name{dmesh_effort}
\alias{dmesh_effort}
\title{Summarize effort to dual mesh}
\usage{
dmesh_effort(
  dmesh,
  obs,
  background,
  adjust = TRUE,
  buffer = NULL,
  nsimeff = 20
)
}
\arguments{
\item{dmesh}{A dual mesh sf object}

\item{obs}{An sf data.frame with observations of the target species}

\item{background}{An sf data.frame with observations of the target group}

\item{adjust}{Whether to adjust effort to be species-specific}

\item{buffer}{A sf polygon to be used as a buffer around locations to prevent extrapolation outside of the species range. Dual mesh cells without any effort outside of this buffer will be assigned an effort value to force model predictions toward 0.}

\item{nsimeff}{Effort value to assign to cells outside of the buffer (\code{integer} representing an umber of background observations).}

\item{\dots}{Arguments passed to \code{inla}}
}
\value{

}
\description{
This function summarizes effort to dual mesh.
}
\details{
Either points or a raster can be used. If a raster, empty cells will be filled with 0.
}
\references{
Simpson, D. Illian, J. B., Lindgren, F. Sørbye, S. H. and Rue, H. 2016. Going off grid: computationally efficient inference for log-Gaussian Cox processes. Biometrika, 103(1): 49-70 \url{https://doi.org/10.1093/biomet/asv064}
}
\keyword{}
