% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmesh_effort.R
\name{dmesh_effort}
\alias{dmesh_effort}
\title{Summarize effort to dual mesh}
\usage{
dmesh_effort(
  dmesh,
  obs,
  background,
  adjust = FALSE,
  buffer = NULL,
  nsimeff = 20
)
}
\arguments{
\item{dmesh}{A dual mesh sf object}

\item{obs}{A sf spatial data.frame with observations of the target species}

\item{background}{A sf data.frame with observations of the target group or a terra raster with the sum of observations of the target group for each pixel.}

\item{adjust}{Whether to adjust effort to be species-specific. Default to \code{FALSE}. If \code{adjust = TRUE}, a column named \"species\" with species name must be present in the \code{background data.frame}. Currently ignored for when a species column is not given in \code{background} or when it is a raster.}

\item{buffer}{A sf polygon to be used as a buffer around locations to prevent extrapolation outside of the species range. Dual mesh cells without any effort outside of this buffer will be assigned an effort value to force model predictions toward 0.}

\item{nsimeff}{Effort value to assign to cells outside of the buffer. An \code{integer} representing a number of background observations.}

\item{\dots}{Arguments passed to \code{inla}}
}
\value{
A list with element effort with a \code{data.frame} summarizing the number of observations and the various effort measures for each cell of the dual mesh.

Depending on the options chosen, the \code{data.frame} will contain some or all of the following:

\itemize{
\item\code{nobs}{ : number of observations}
\item\code{nbackground}{ : number of background observations from the target group}
\item\code{npres}{ : whether the species is present in a cell or not (1 = present, 0 = absent)}
\item\code{nsp}{ : number of species in a cell}
\item\code{nbackgroundwithbuff}{ : nbackground to which fictious observations have been added using the effort buffer}
\item\code{nbackgroundspadjusted}{ : nbackground with species-specific adjustment}
\item\code{nbackgroundspadjustedwithbuff}{ : nbackground with species-specific adjustment and to which fictious observations have been added using the effort buffer}
}
}
\description{
This function summarizes effort to dual mesh.
}
\details{
Either an \code{sf} spatial object with points or a raster can be used. If a raster, empty cells will be filled with 0.
}
\references{
Simpson, D. Illian, J. B., Lindgren, F. Sørbye, S. H. and Rue, H. 2016. Going off grid: computationally efficient inference for log-Gaussian Cox processes. Biometrika, 103(1): 49-70 \url{https://doi.org/10.1093/biomet/asv064}
}
