% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmesh_weights.R
\name{dmesh_weights}
\alias{dmesh_weights}
\title{Dual Mesh Weights}
\usage{
dmesh_weights(dmesh, region)
}
\arguments{
\item{dmesh}{Dual mesh}

\item{region}{A region}
}
\value{
A list with element weights containing a vector of areas for each cell within the region area
}
\description{
Computes the dual mesh weights for the integration of the point process
}
\details{
none
}
\examples{
none

}
\references{
none
}
