% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewlgcp.R
\encoding{UTF-8}
\name{ewlgcp}
\alias{ewlgcp}
\title{Effort-Weighted Log Gaussian Cox Process SDM}
\usage{
ewlgcp(
  formula,
  dmesh,
  effort = TRUE,
  adjust = TRUE,
  buffer = TRUE,
  orthogonal = TRUE,
  prior.beta = NULL,
  prior.range = c(50, 0.1),
  prior.sigma = c(1, 0.1),
  smooth = 3/2,
  ...
)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x1 + x2 + ...}.}

\item{dmesh}{A list with elements produced by the different \code{dmesh_} functions.}

\item{effort}{Logical. Whether to adjust the model for effort. Default \code{TRUE}.}

\item{adjust}{Logical. Whether to adjust the effort for being species specific. Default \code{TRUE}.}

\item{buffer}{Logical. Whether to add the effort buffer to help reducing prediction outside of the species range. Default \code{TRUE}.}

\item{orthogonal}{Logical. Whether to make the spatial field orthogonal to the predictors? Default \code{TRUE}.}

\item{prior.beta}{Normal priors for the betas of the fixed effects coefficients as required by \code{INLA}. Default is \code{list(prec=list(default=1/(1)^2,Intercept=1/(20)^2),mean=list(default=0,Intercept=0))} which means a prior with \code{mean = 0} and \code{sd = 1} for all coefficients and a prior with \code{mean = 0} and \code{sd = 20} for the model intercept. See \code{\link{?control.fixed}}.}

\item{prior.range}{Penalized complexity prior for the range of the spatial field. A vector of length two giving the probability that the range is inferior to a given value. The default is \code{prior.range = c(50, 0.01)} which represents a 1\% chance that the range is inferior to 50 (in the units of the crs used).}

\item{prior.sigma}{Penalized complexity prior for the standard deviation (sd) of the spatial field. A vector of length two giving the probability that the sd is superior to a given value. The default is \code{prior.sigma = c(1, 0.01)} which represents a 1\% chance that the range is superior to 1.}

\item{smooth}{x}

\item{\dots}{Further arguments to pass to \code{inla}}
}
\value{
A model of class \code{INLA}.
}
\description{
Runs the effort-weighted Log Gaussian Cox Process species distribution model
}
\details{
none
}
\examples{
none



}
\references{
Simpson, D. Illian, J. B., Lindgren, F. S\u00f8rbye, S. H. and Rue, H. 2016. Going off grid: computationally efficient inference for log-Gaussian Cox processes. Biometrika, 103(1): 49-70 \url{https://doi.org/10.1093/biomet/asv064}

Fuglstad, G.-A., Simpson, D., Lindgren, F. & Rue, H. 2019 Constructing Priors that Penalize the Complexity of Gaussian Random Fields. Journal of the American Statistical Association, 114(525): 445-452 \url{https://doi.org/10.1080/01621459.2017.1415907}
}
